+++ RevitAddin.FamilyLoader/Models/FamilyLoaderModel.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace RevitAddin.FamilyLoader.Models
+{
+    public class FamilyLoaderModel
+    {
+        public string Name { get; set; }
+        public string Symbol { get; set; }
+        public object Image { get; set; }
+        public string Path { get; set; }
+    }
+}