+++ RevitAddin.FamilyLoader/Revit/App.cs
@@ -1,6 +1,7 @@
 using Autodesk.Revit.DB;
 using Autodesk.Revit.UI;
 using ricaun.Revit.UI;
+using ricaun.Revit.UI.Tasks;
 using System;
 
 namespace RevitAddin.FamilyLoader.Revit
@@ -9,17 +10,26 @@ namespace RevitAddin.FamilyLoader.Revit
     public class App : IExternalApplication
     {
         private RibbonPanel ribbonPanel;
+        private static RevitTaskService RevitTaskService;
+        public static IRevitTask RevitTask => RevitTaskService;
         public Result OnStartup(UIControlledApplication application)
         {
             ribbonPanel = application.CreatePanel("FamilyLoader");
-            ribbonPanel.CreatePushButton<Commands.Command>("Load\rFamily")
+            ribbonPanel.CreatePushButton<Commands.CommandView>("Load\rFamily")
                 .SetLargeImage("Resources/Revit.ico");
+
+            RevitTaskService = new RevitTaskService(application);
+            RevitTaskService.Initialize();
+
             return Result.Succeeded;
         }
 
         public Result OnShutdown(UIControlledApplication application)
         {
             ribbonPanel?.Remove();
+
+            RevitTaskService.Dispose();
+
             return Result.Succeeded;
         }
     }