+++ RevitAddin.FamilyLoader/Revit/Commands/Command.cs
@@ -36,4 +36,5 @@ namespace RevitAddin.FamilyLoader.Revit.Commands
             return Result.Succeeded;
         }
     }
+
 }